+++ packages/configs/eslint-config-mdx/src/remarkrc.ts
@@ -1,0 +1,28 @@
+import remarkLintCodeFormat from "./remark-lint-code-format.js"
+import remarkLintMdxJsxFormat from "./remark-lint-mdx-jsx-format.js"
+import remarkPreserveAlertMarkers from "./remark-preserve-alert-markers.js"
+
+const remarkConfig = {
+  plugins: [
+    "remark-frontmatter",
+    "remark-lint-fenced-code-flag",
+    "remark-lint-final-newline",
+    "remark-lint-heading-increment",
+    "remark-lint-list-item-content-indent",
+    ["remark-lint-list-item-indent", "one"],
+    ["remark-lint-maximum-heading-length", {size: 80}],
+    "remark-lint-no-consecutive-blank-lines",
+    "remark-lint-no-duplicate-headings-in-section",
+    ["remark-lint-no-heading-punctuation", ",.:;"],
+    ["remark-lint-unordered-list-marker-style", "-"],
+    remarkLintCodeFormat,
+    remarkLintMdxJsxFormat,
+    remarkPreserveAlertMarkers,
+  ],
+  settings: {
+    bullet: "-",
+    rule: "-",
+  },
+}
+
+export default remarkConfig